+++ packages/oidc/lib/src/_exports.dart
@@ -1,3 +1,2 @@
 export 'facade.dart';
 export 'managers/_exports.dart';
-export 'models/_exports.dart';
