+++ packages/oidc_core/lib/src/managers/_exports.dart
@@ -1,1 +1,2 @@
 export 'token_events.dart';
+export 'user_manager_base.dart';
