+++ packages/oidc_core/lib/src/models/_exports.dart
@@ -1,6 +1,8 @@
 export '../constants.dart';
 export 'client_auth.dart';
 export 'error.dart';
+export 'events/_exports.dart';
 export 'exception.dart';
+export 'settings/_exports.dart';
 export 'state/_exports.dart';
 export 'user.dart';
