+++ packages/oidc_core/lib/src/models/events/_exports.dart
@@ -1,0 +1,2 @@
+export 'event.dart';
+export 'pre_logout_event.dart';
