+++ packages/oidc_core/lib/src/models/settings/_exports.dart
@@ -1,3 +1,4 @@
-export 'event.dart';
+export 'front_channel_request_options.dart';
 export 'id_token_verification_options.dart';
+export 'platform_options.dart';
 export 'user_manager_settings.dart';
