+++ packages/oidc_core/lib/src/models/settings/front_channel_request_options.dart
