+++ packages/oidc_core/lib/src/models/settings/id_token_verification_options.dart
