+++ packages/oidc_core/lib/src/models/settings/platform_options.dart
