+++ packages/oidc_core/pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   logging: ^1.2.0
   clock: ^1.1.1
   copy_with_extension: ^5.0.4
+  rxdart: ^0.27.7
 
 dev_dependencies:
   build_runner: ^2.4.6
