+++ packages/oidc_desktop/test/oidc_desktop_test.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:logging/src/logger.dart';
+import 'package:oidc_core/oidc_core.dart';
 import 'package:oidc_desktop/oidc_desktop.dart';
 import 'package:oidc_platform_interface/oidc_platform_interface.dart';
