+++ packages/oidc_macos/lib/oidc_macos.dart
@@ -1,5 +1,5 @@
+import 'package:oidc_core/oidc_core.dart';
 import 'package:oidc_flutter_appauth/oidc_flutter_appauth.dart';
-
 import 'package:oidc_platform_interface/oidc_platform_interface.dart';
 
 /// The MacOS implementation of [OidcPlatform].
