+++ packages/oidc_platform_interface/lib/oidc_platform_interface.dart
@@ -1,2 +1,1 @@
-export 'src/models/_exports.dart';
 export 'src/platform.dart';
