+++ packages/oidc_web_core/example/README.md
@@ -1,0 +1,24 @@
+A bare-bones Dart web app.
+
+Uses [`package:web`](https://pub.dev/packages/web)
+to interop with JS and the DOM.
+
+## Running and building
+
+To run the app,
+activate and use [`package:webdev`](https://dart.dev/tools/webdev):
+
+```
+dart pub global activate webdev
+webdev serve
+```
+
+To build a production version ready for deployment,
+use the `webdev build` command:
+
+```
+webdev build
+```
+
+To learn how to interop with web APIs and other JS libraries,
+check out https://dart.dev/interop/js-interop.
