+++ packages/oidc_web_core/example/pubspec.yaml
@@ -1,0 +1,20 @@
+name: oidc_web_core_example
+description: An absolute bare-bones web app that uses oidc protocol.
+version: 1.0.0
+publish_to: none
+# repository: https://github.com/my_org/my_repo
+
+environment:
+  sdk: ^3.4.3
+
+# Add regular dependencies here.
+dependencies:
+  web: ^0.5.1
+  oidc_core: ^0.6.3
+  oidc_web_core:
+    path: ../
+
+dev_dependencies:
+  build_runner: ^2.4.8
+  build_web_compilers: ^4.0.9
+  lints: ^3.0.0
