+++ packages/oidc_web_core/example/web/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+
+<html>
+
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta name="scaffolded-by" content="https://github.com/dart-lang/sdk">
+  <title>oidc_web_core example</title>
+  <link rel="stylesheet" href="styles.css">
+  <script defer src="main.dart.js"></script>
+</head>
+
+<body>
+
+  <div id="output"></div>
+  <button id="loginButton">Login with duende</button>
+  <div id="userOutput"></div>
+
+
+</body>
+
+</html>
