+++ packages/oidc_web_core/lib/oidc_web_core.dart
@@ -1,5 +1,6 @@
 /// Core web package for oidc
 library;
 
-export 'src/models/_exports.dart';
 export 'src/oidc_web_core.dart';
+export 'src/oidc_web_store.dart';
+export 'src/user_manager_web.dart';
