--- packages/oidc_web_store/CHANGELOG.md
@@ -1,3 +1,0 @@
-# 0.1.0+1
-
-- feat: initial commit 🎉
