--- packages/oidc_web_store/lib/oidc_web_store.dart
@@ -1,4 +1,0 @@
-/// an implementation of OidcStore for web that doesn't depend on flutter
-library;
-
-export 'src/oidc_web_store.dart';
