--- packages/oidc_web_store/pubspec.yaml
@@ -1,17 +1,0 @@
-name: oidc_web_store
-description: an implementation of OidcStore for web that doesn't depend on flutter
-version: 0.1.0+1
-
-
-environment:
-  sdk: ^3.4.0
-
-dependencies: 
-  collection: ^1.18.0
-  oidc_core: ^0.6.3
-  web: ^0.5.1
-  
-dev_dependencies:
-  mocktail: ^1.0.3
-  test: ^1.25.2
-  very_good_analysis: ^5.1.0
