--- packages/oidc_web_store/test/src/oidc_web_store_test.dart
@@ -1,11 +1,0 @@
-// ignore_for_file: prefer_const_constructors
-import 'package:oidc_web_store/oidc_web_store.dart';
-import 'package:test/test.dart';
-
-void main() {
-  group('OidcWebStore', () {
-    test('can be instantiated', () {
-      expect(OidcWebStore(), isNotNull);
-    });
-  });
-}
