+++ packages/oidc_windows/test/oidc_windows_test.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:oidc_core/oidc_core.dart';
 import 'package:oidc_platform_interface/oidc_platform_interface.dart';
 import 'package:oidc_windows/oidc_windows.dart';
